requirements.txt
# ── requirements.txt ─────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
supabase==2.7.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
redis==5.1.0
slowapi==0.1.9
twilio==9.3.0
firebase-admin==6.5.0
celery==5.4.0
flower==2.0.0
sentry-sdk[fastapi]==2.15.0
structlog==24.4.0
